pyproject.toml
[project]
name = "fastqueue"
version = "0.1.0"
description = "A fast, lightweight, open-source SQS alternative — deployable locally or in production, powered by in-memory queues and a clean REST API."
authors = [
    {name = "Priyanshu Panwar",email = "priyanshu009ch@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "typer (>=0.15.4,<0.16.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-env = "^1.1.5"
